Create Auto_f2 V3.0.bat
Auto_f2 V3.0.bat is a sequel to AIO_f1 V2.0.bat. AIO (All In One)_f1(from V1.0) V2.0 (version 2.0) could only retrieve and send the coordinates. Auto (Automatic execute)_f2 (from V2.0) V3.0 (version 3.0).bat moves itself to the startup folder, so that the coordinates are shared every time the computer is started. However, if it is already there, it only sends the coordinates. In doing so, it prints "systemically relevant" messages in the cmd field so as not to arouse suspicion.